BACKPORT: ASoC: intel: sof_sdw: Add simple DAI link creation helper
The code contains a fair amount of state tracking and one part of that
is keeping track of which entry in the large global cpus
snd_soc_dai_link_component array is currently in use. Add a helper
function to allocate a simple DAI link, this simplifies the
code slightly and moves us in the direction of eliminating the need for
the large global cpus array. This does slightly increase the number of
allocations done, but this is probe time and the code already does a
large number of allocations so this increase is small over all.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Link: https://lore.kernel.org/r/20230915075611.1619548-3-yung-chuan.liao@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit b359760d95eecaabd081c1c2cd58e0a15fe5a68c)

Conflicts:
   sound/soc/intel/boards/sof_sdw.c

BUG=b:315885644
TEST=Success to build kernel

Change-Id: Ide13f3ff4aab2dc6e029d44ec427fc59f72d7d4a
Signed-off-by: Brent Lu <brent.lu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5553113
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Terry Cheong <htcheong@chromium.org>