CHROMIUM: drm/mediatek: Check allocation size
Add an extra check in the MTK gpu driver validating the size of
requested allocation. If the size is == 0 the driver tries to allocate
it resulting in nullptr which is further dereferenced. This leads to
kernel panic which should never occur. Check the requested size and
in case of invalidity log the event and return the appropriate error.

BUG=b:299095063
TEST=android.media.cts.EncodeDecodeTest
UPSTREAM-TASK=b:302067535

Signed-off-by: Hubert Mazur <hmazur@google.com>
Change-Id: If3d4c3bc82234f6b7f191d3850362648a88e5ecc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4876640
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>