CHROMIUM: sched/core: sched_setattr_pi_nocheck
__sched_setscheduler takes an argument 'pi' so as to allow its usage in
interrupt context when PI is not used. But this is not exported and
cannot be used outside of scheduler code. sched_setattr_nocheck is
exported but it doesn't allow that flexibility.

Introduce sched_setattr_pi_nocheck to allow for the flexibility to
call from interrupt context.

UPSTREAM-TASK=b:303645537
BUG=b:262267726
TEST=boot

Co-developed-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Vineeth Pillai (Google) <vineeth@bitbyteword.org>

Change-Id: I8bd03ee796df29a641be762188a176758330377c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5249682
Tested-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-by: Joel Fernandes <joelaf@google.com>
Commit-Queue: Vineeth Pillai <vineethrp@google.com>