FROMGIT: drm/i915/hdcp: Add more conditions to enable hdcp
When we dock a monitor we end up with a enable and disable connector
cycle but if hdcp content is running we get the userspace in
enabled state and driver maintaining a undesired state which causes
the content to stop playing and we only enable hdcp if the userspace
state in desired. This patch fixes that.

--v2
-Move code to intel_hdcp [Jani]

Signed-off-by: Suraj Kandpal <suraj.kandpal@intel.com>
Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Signed-off-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20231026121139.987437-4-suraj.kandpal@intel.com
(cherry picked from commit 3b9bbd79627043a9fa9dd5b01bb29882663976e0 https://anongit.freedesktop.org/git/drm-intel drm-intel-next)

BUG=b:289984277
BUG=b:247941792
TEST= Test docked mode video playback on netflix with HDCP2.2 monitor.

Change-Id: If8d85809b0247a34403528c279670f03e656f910
Signed-off-by: FNU VISHWANATHA <fnu.vishwanatha@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4995843
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>