FROMGIT: kvm: x86: paravirt sched framework
Implement the basic framework needed for guest/host communication to
facilitate paravirt sched:
 - A kvm MSR that guest uses to provide the GPA of shared memory for
   communicating the scheduling information between host and guest.

   wrmsr(0) disables the feature. wrmsr(valid_gpa) enables the feature
   and uses the gpa for further communication.

 - Add a new cpuid feature flag for the host to advertise the feature
   to the guest.

When the guest kernel is about to run a critical or latency sensitive
workload, it can request the hypervisor to boost the priority of the
vcpu thread. Similarly, guest kernel can request to unboost when the
vcpu switches to a normal workload. Guest kernel can also share the
priority attributes of the task that it is going to schedule and host
can adjust the priority of vcpu thread acordingly.

When a guest determines that it needs a boost, it need not immediately
request a synchronous boost as it is already running at that moment.
Synchronous request is detrimental because it incurs a VMEXIT. Rather,
let the guest note down its request on a shared memory and the host can
check this request on next VMEXIT and boost if needed.

Nmis, irqs, softirqs and preemption disabled are considered latency
sensitive critical sections in kernel and are candidates for boosting.

Co-developed-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Signed-off-by: Vineeth Pillai (Google) <vineeth@bitbyteword.org>

Change-Id: I4a63df3d12bec9026fcb7c82f82b4bdf02cbb0b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5249683
Reviewed-by: Joel Fernandes <joelaf@google.com>
Commit-Queue: Vineeth Pillai <vineethrp@google.com>
Tested-by: Vineeth Pillai <vineethrp@google.com>
(cherry picked from commit 6511976fd6ddbf3f203722d8d76584e746ee698f)