FROMGIT: drm/i915: Fix ADL+ tiled plane stride when the POT stride is smaller than the original
plane_view_scanout_stride() currently assumes that we had to pad the
mapping stride with dummy pages in order to align it. But that is not
the case if the original fb stride exceeds the aligned stride used
to populate the remapped view, which is calculated from the user
specified framebuffer width rather than the user specified framebuffer
stride.

Ignore the original fb stride in this case and just stick to the POT
aligned stride. Getting this wrong will cause the plane to fetch the
wrong data, and can lead to fault errors if the page tables at the
bogus location aren't even populated.

TODO: figure out if this is OK for CCS, or if we should instead increase
the width of the view to cover the entire user specified fb stride
instead...

Cc: Imre Deak <imre.deak@intel.com>
Cc: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Change-Id: Ibaf1bf395160d65d784553384841dcc0433fd01e
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20231204202443.31247-1-ville.syrjala@linux.intel.com
Reviewed-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
(cherry picked from commit 01a39f1c4f1220a4e6a25729fae87ff5794cbc52
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:279521187, b:289273022
TEST=No DMAR error message while running test application in issue #279521187.

Change-Id: I711056202885b3bcb443ee6444e34566b9998b7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5091331
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Sean Paul <sean@poorly.run>