CHROMIUM: sched/eevdf: Adding fair_eevdf.c
The objective of this change series is to be able to backport and
experiment with EEVDF. In order to not mess up any backports from
stable, this does not modify the fair.c file, but instead makes a
copy of it as fair_eevdf.c. Then the backport of the EEVDF scheduler
will update only this copy allowing normal backports from stable to
not conflict with fair.c.

BUG=b:308209790
TEST=Manual tested on DUT
UPSTREAM-TASK=b:324600165

Signed-off-by: Youssef Esmat <youssefesmat@chromium.org>

Change-Id: Ice282f1bd53c49866a5f9c2d0c7ca45a31f2f679
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5089890
Reviewed-by: Vineeth Pillai <vineethrp@google.com>
Commit-Queue: Youssef Esmat <youssefesmat@chromium.org>
Tested-by: Vineeth Pillai <vineethrp@google.com>