BACKPORT: sched/fair: Add cfs_rq::avg_vruntime
In order to move to an eligibility based scheduling policy, we need
to have a better approximation of the ideal scheduler.

Specifically, for a virtual time weighted fair queueing based
scheduler the ideal scheduler will be the weighted average of the
individual virtual runtimes (math in the comment).

As such, compute the weighted average to approximate the ideal
scheduler -- note that the approximation is in the individual task
behaviour, which isn't strictly conformant.

Specifically consider adding a task with a vruntime left of center, in
this case the average will move backwards in time -- something the
ideal scheduler would of course never do.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lore.kernel.org/r/20230531124603.654144274@infradead.org

BUG=b:308209790
TEST=Manual tested on DUT

Signed-off-by: Youssef Esmat <youssefesmat@chromium.org>

Change-Id: I97a2b8d1b05992742bb878a7d6e0124dbcf11d27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5089891
Reviewed-by: Vineeth Pillai <vineethrp@google.com>
Tested-by: Vineeth Pillai <vineethrp@google.com>
Commit-Queue: Youssef Esmat <youssefesmat@chromium.org>