CHROMIUM: config: arm64: Rename chromiumos-arm64.flavour.config
chromiumos-<arch>.flavour.config and common.config can be ambiguous
sometimes if they are only referenced by their names.

Append `-generic` to the former and hopefully make it clearer that it's
for building a generic kernel, not for putting common kernel configs
that will be applied to all platforms.

BUG=b:317315065
TEST=emerge-arm64-generic chromeos-kernel-$version

Cq-Depend: chromium:5131878
Change-Id: I9303f8d0913dec8b496d89cac22cb2e1336b6705
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5133978
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>