Revert "CHROMIUM: hung_task: dump all UNINTERUPTIBLE tasks"
This reverts commit 21e867019c11a67e6aa3ea3d302deeedb87c8b40.

The downstream commit doesn't really make so much sense after upstream
commit 401c636a0eeb ("kernel/hung_task.c: show all hung tasks before
panic").

Specifically, the idea of the downstream commit was that it's possible
a task is blocked because it is waiting for a resource held by another
task that itself was blocked. Back before we showed all hung tasks we
wouldn't see its stack. Now, we will. That means the extra dump that
the downstream patch is doing is just showing the same stacks again
and that could blow out older kernel log messages that are important.

BUG=None
TEST=Run several `echo HUNG_TASK > /sys/kernel/debug/provoke-crash/DIRECT &`

Change-Id: Iba2b414406097907d43bf1815c8ff24257af7bac
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5208312
Reviewed-by: Guenter Roeck <groeck@chromium.org>