CHROMIUM: drm/bridge: parade-ps8640: Reset the bridge if the HPD didn't get asserted
There are reports from the field of some devices with ps8640 failing
to bring up their panel. Specifically in the logs we see:

  panel-simple-dp-aux aux-1-0008: Prepare timeout after 5 tries
  panel-simple-dp-aux aux-1-0008: Couldn't power on panel to read EDID: -110
  panel-simple-dp-aux aux-1-0008: error -ETIMEDOUT: Couldn't detect panel nor find a fallback

As can be seen from the above, the eDP panel driver tries to power on
the panel several times since some panels have been seen to fail to
come up reliably. However, between retries the ps8640 bridge chip
isn't power cycled/reset.

Normally one shouldn't need to power cycle the bridge chip between
retries of the panel. The bridge chip's only real involvement here
should be its ability to read the HPD pin and, one would think, that
couldn't get messed up. However, we read ps8640's HPD not through a
hardware register but by a location signaled by the firmware running
on ps8640. It's possible that this firmware could get confused and so
resetting it when we fail to see HPD isn't a terrible idea.

I don't have any way to reproduce this problem myself, but I simulated
the timeout to make sure that the reset code ran and see that things
still initialized cleanly.

NOTE: this will make it even _slower_ if userspace tries to access the
panel while it's off through the exposed i2c bus or the DP AUX char
device. Presumably we'd want to land the patch ("drm/dp: Don't attempt
AUX transfers when eDP panels are not powered") [1] before this one to
avoid that problem.

[1] https://lore.kernel.org/r/20240202141109.1.I24277520ac754ea538c9b14578edc94e1df11b48@changeid

CHROMIUM NOTE: proposal is to land this patch locally to get some
confidence that it fixes the problem in the field, then either drop it
or send it upstream.

UPSTREAM-TASK=b:323905518
BUG=b:304370964
TEST=Simulate timeout

Disallow-Recycled-Builds: test-failures
Change-Id: I72bcbbe65e65162f78b24d3e93089d1a8565d5fc
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5264198
Reviewed-by: Prahlad Kilambi <prahladk@google.com>