CHROMIUM: arm64: dts: mt8186: Modify sbs battery reg to 0x0f
During the process of battery charging and discharging,
the display power will jump. This is because when reading the power,
the display power is calculated in watts due to unsynchronized
reading and writing.
Use virtual battery 0xf to communicate with EC can solve this.

BUG=b:319031888
TEST=check dispaly battery

Change-Id: Ide4e6bd915ca890836f664ff65e50191ef5af1f8
Signed-off-by: Zhengqiao Xia <xiazhengqiao@huaqin.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5266218
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Jiahong Wei <weijiahong@huaqin.corp-partner.google.com>