BACKPORT: FROMGIT: arm64/sve: Lower the maximum allocation for the SVE ptrace regset
Doug Anderson observed that ChromeOS crashes are being reported which
include failing allocations of order 7 during core dumps due to ptrace
allocating storage for regsets:

  chrome: page allocation failure: order:7,
          mode:0x40dc0(GFP_KERNEL|__GFP_COMP|__GFP_ZERO),
          nodemask=(null),cpuset=urgent,mems_allowed=0
   ...
  regset_get_alloc+0x1c/0x28
  elf_core_dump+0x3d8/0xd8c
  do_coredump+0xeb8/0x1378

with further investigation showing that this is:

   [   66.957385] DOUG: Allocating 279584 bytes

which is the maximum size of the SVE regset. As Doug observes it is not
entirely surprising that such a large allocation of contiguous memory might
fail on a long running system.

The SVE regset is currently sized to hold SVE registers with a VQ of
SVE_VQ_MAX which is 512, substantially more than the architectural maximum
of 16 which we might see even in a system emulating the limits of the
architecture. Since we don't expose the size we tell the regset core
externally let's define ARCH_SVE_VQ_MAX with the actual architectural
maximum and use that for the regset, we'll still overallocate most of the
time but much less so which will be helpful even if the core is fixed to
not require contiguous allocations.

Specify ARCH_SVE_VQ_MAX in terms of the maximum value that can be written
into ZCR_ELx.LEN (where this is set in the hardware). For consistency
update the maximum SME vector length to be specified in the same style
while we are at it.

We could also teach the ptrace core about runtime discoverable regset sizes
but that would be a more invasive change and this is being observed in
practical systems.

Reported-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Link: https://lore.kernel.org/r/20240213-arm64-sve-ptrace-regset-size-v2-1-c7600ca74b9b@kernel.org
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit 2813926261e436d33bc74486b51cce60b76edf78
 git://git.kernel.org/pub/scm/linux/kernel/git/arm64/linux for-next/fixes)

Conflicts:
   arch/arm64/include/asm/fpsimd.h
Conflict is because we are lacking commit d158a0608eb8 ("arm64/sme:
More sensibly define the size for the ZA register set") which added
the "#define SME_VQ_MAX 16" and the comment around it. We don't need
that patch, though, since we don't have the commits it fixes, AKA
commit 776b4a1cf364 ("arm64/sme: Add ptrace support for ZA") and
commit e12310a0d30f ("arm64/sme: Implement ptrace support for
streaming mode SVE registers"). Picking the series including those
would be over the top. It should also be noted that the commit we're
cherry-picking is a no-op without commit f171f9e4097d ("arm64/fp: Make
SVE and SME length register definition match architecture"). To avoid
needing to pick the extra dependency (with yet more conflicts), we'll
just replace the #define with the result of 16 which matches what it
evaluates to and also what was in v1.

BUG=b:323363415
TEST=Printouts confirm alloc is smaller

Change-Id: Ib49e37bde376507c42ff93628d3397333c2387ec
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5300889
Reviewed-by: Rob Clark <robdclark@chromium.org>