CHROMIUM: Reduce DPM watchdog timeout to 10 seconds.
DPM_WATCHDOG is not catching any issues as EC identifies hangs and
resets in 20 seconds. Reduce DPM_WATCHDOG timeout to 10 seconds. This is under all the 
suspend/resume time upperbound(~8s) for all chromeos in crosbolt.

BUG=b:327508225
TEST=./chromeos/scripts/kernelconfig checkconfig
On redrix, verified by simulating dpm hang and dpm watch dog catches it in 10 seconds before EC
identifies it and causes a single reboot with console-ramoops.

Change-Id: Id2e2dcb8ac6a29493a53198c22ccc001437bc5de
Signed-off-by: Lalith Rajendran <lalithkraj@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5341615
Reviewed-by: Tim Van Patten <timvp@google.com>