FROMLIST: sched/deadline: Always start a new period if CFS exceeded DL runtime
We believe that this is the right thing to do. The unit test
(cs_dlserver_test) also agrees. If we let the CFS run without starting a
new period, while the server is regularly throttled, then the test fails
because CFS does not appear to get enough bandwidth.

Intuitively, this makes sense to do as well. If CFS used up all the CFS
bandwidth, while the DL server was in a throttled state, it got the
bandwidth it wanted and some. Now, we can start all over from scratch to
guarantee it a minimum bandwidth.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>

(am from https://lore.kernel.org/lkml/20240313012451.1693807-16-joel@joelfernandes.org/)

BUG=b:307273029
TEST=booted

Change-Id: I3a700edfbf586f7f38bfd0a7e95d6b2bd37af53d
Signed-off-by: Suleiman Souhlal <suleiman@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5377583
Reviewed-by: Vineeth Pillai <vineethrp@google.com>