CHROMIUM: media: camx: rework fault handler
camx fault handler does two things from atomic context:
1) does GFP_KERNEL allocation
2) takes a mutex

Remove that allocation, it seems that we can use just a stack
variable. We also don't seem to be using the work queue the
way it is supposed to be (for deferred execution), but instead
it handles the fault in-inline. So we don't need the work-queue,
its list and mutex.

UPSTREAM-TASK=b:274966069
BUG=b:331961179
TEST=CTS on coachz

Change-Id: Ifb140f731d142505031c3214764e8fcaa81f28e7
Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5409738
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5471546