CHROMIUM: media: camx: disable IRQs when acquire clients_lock
IRQs need to be disabled every-time we take clients_lock.
The reason being is that PF is handled in the IRQ context,
which calls cam_cdm_notify_clients() and attempts to take
clients_lock, so we cannot let PF IRQ to occur when we
already hold clients_lock.

UPSTREAM-TASK=b:274966069
BUG=b:331961179
TEST=CTS on wormdingler

Change-Id: Id0f9e067957d0fe083a555328fa7c72da19527db
Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5465219
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5471557