CHROMIUM: img-rogue: FW command size mismatch
The FW commands passed in the KickTA3D bridge have a shared
part that is filled in by KM. The FW command size parameter
passed from UM could be set to smaller than size of shared
struct CMDTA3D_SHARED. This can lead to size mismatch.

Ensure size of the commands: TA, 3D and 3DPR are at least
the size of the struct that KM writes.

IMG Ref: PP159407
BUG=b:332938679
TEST=Boot to UI, PoC and graphics.WebGLAquarium.1000_fishes

Change-Id: I2d9073825448834dff4daf91720538b3c5055345
Signed-off-by: Faijur Rahman <faijur.rahman@imgtec.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5545626
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Tested-by: Hsin-Te Yuan <yuanhsinte@chromium.org>