Revert "security: selinux: Guard KernelSU changes"
The changes on guarding KernelSU in the SELinux policy picked from a 4.14 kernel actually causes problems
on mounting root access with 5.15 kernels, which results in an unknown statement and the root access' being
partly unusable after mounting. Currently I'm not so clear about it so I have to revert the commit.

This reverts commit 2ee0d164e570f545d55049ddd32b17f5a7a5fa64.

Signed-off-by: crrashh1542 <kotsuki@crrashh.com>