CHROMIUM: ASoC: mt6358: Add "Dmic Mode Switch" kcontrol for switch DMIC mode.
There are two hardware connection methods for DMICs on the MT6358. In cases where more than two
DMICs are used, we need to time-multiplex these DMICs. Therefore, we need to dynamically switch
the modes of these DMICs based on the actual usage scenarios.

           ---- DMIC1
AU_VIN0 ---
           ---- DMIC2

AU_VIN2 --- ----DMIC3

When we want to use DMIC1/2, configure it to one-wire mode. When we want to use DMIC1/3, configure
it to two-wire mode.

UPSTREAM-TASK=b:336915002
BUG=b:336915002
TEST= boot to shell and verify ok with amixer.

Change-Id: I4183c73304901f95997a33c8b9c94631a33223f8
Signed-off-by: Jiaxin Yu <jiaxin.yu@mediatek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5584176
Commit-Queue: Baili Deng <bailideng@google.com>
Reviewed-by: Baili Deng <bailideng@google.com>