CHROMIUM: arm64: dts: mt8186: Modify sbs battery reg to 0x0f
During the process of battery charging and discharging,
the display power will jump. This is because when reading
the power, the display power is calculated in watts due to
unsynchronized reading and writing. Use virtual battery 0xf
to communicate with EC can solve this.

BUG=b:344634088
TEST=check battery charging and discharge function
UPSTREAM-TASK=b:345520795

Change-Id: Iea63b60c612d6c795b3934ad31525e84f65639fd
Signed-off-by: Asher Huang <asher_huang@pegatron.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5607102
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Knox Chiou <knoxchiou@chromium.org>